+++ noodles-bam/CHANGELOG.md
@@ -18,6 +18,8 @@
 
   * bam/async/writer/record: Calculate bin number for record.
 
+  * bam/bai/index/reference_sequence/builder: Calculate bin number for record.
+
   * bam/record/builder: Fix default read name to have no NUL terminator.
 
   * bam/writer/record: Calculate bin number for record.